Fixed a number of issues: PHS mass is now correctly added to the top and bottom z-surfaces The disclination hopping for the site-centered plaquette is correct now Added NNN hoppings across the plaquette-centered disclination
Also made some improvements to code readability. Added a function that creates a hopping matrix across the disclination in the same way as is done for x- and y-hopping

Still need to restructure the code so that the half/other model parameter works nicely like the spin parameter

Still need to combine the site- and plaquette-centered disclination codes into one. This will also require rewriting anything that reads or writes data